# Conflicts:
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/availableproductspage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/categorypage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/outletpage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/productpage.js
#	be/src/Project/Web/Buyer/Buyer.Web/wwwroot/dist/js/searchproductspage.js

+++ be/src/Project/Services/Inventory/Inventory.Api/v1/ResponseModels/InventoryDetailsResponseModel.cs
@@ -10,9 +10,9 @@ namespace Inventory.Api.v1.ResponseModels
         public string ProductSku { get; set; }
         public Guid? WarehouseId { get; set; }
         public string WarehouseName { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Ean { get; set; }
-        public int? AvailableQuantity { get; set; }
+        public double? AvailableQuantity { get; set; }
         public int? RestockableInDays { get; set; }
         public DateTime? ExpectedDelivery { get; set; }
         public DateTime LastModifiedDate { get; set; }
